build: split cryptpilot-fde into host and guest RPM packages
Split the monolithic cryptpilot-fde binary and RPM package into two
separate packages to reduce dependencies in guest disk images:

- cryptpilot-fde-host: disk image conversion (cryptpilot-convert),
  enhancement (cryptpilot-enhance), reference value computation
  (show-reference-value), and config management (config). Pulls in
  all provider features from cryptpilot-core.

- cryptpilot-fde-guest: boot-service binary for initrd decryption
  (dm-crypt, dm-verity, LVM, overlayfs). Minimal dependencies,
  no qemu-img/libguestfs-tools-c/grub2-tools.

Changes:
- Replace single main.rs with bin/host.rs and bin/guest.rs using
  separate CLI structs (Cli and GuestCli)
- Move shared types (metadata, initrd_state) out of guest-only
  boot_service module so both binaries can access them
- Update dracut module-setup.sh and systemd service files to
  reference cryptpilot-fde-guest
- Update cryptpilot-convert.sh to install only cryptpilot-fde-guest
  in the target guest image
- Update Debian packaging (control, rules) accordingly
